Fixed state and renamed code
